+++ MOP/src/Common/CompatibilityManager.cs
@@ -29,13 +29,11 @@ namespace MOP.Common
         // CarryMore
         // https://www.racedepartment.com/downloads/carry-more-backpack-alternative.22396/
         public static bool CarryMore { get; private set; }
-        public static bool CarryEvenMore { get; private set; }
         public readonly Vector3 CarryMoreTempPosition = new Vector3(0.0f, -1000.0f, 0.0f);
 
         public CompatibilityManager()
         {
             CarryMore = ModLoader.GetMod("CarryMore") != null;
-            CarryEvenMore = ModLoader.GetMod("CarryEvenMore") != null;
         }
     }
 }